Simplified solution by getting rid of event again and creating ScriptingContainers in Observer before ResourceProvider gets called.
